move app to `module/`
